Move agent-related code into new root `agent` module
Webserver code that executes the agent remains in `webserver::answer`.
